+++ java-manta-client/src/test/java/com/joyent/manta/client/jobs/MantaJobPhaseTest.java
@@ -1,8 +1,9 @@
 /**
  * Copyright (c) 2015, Joyent, Inc. All rights reserved.
  */
-package com.joyent.manta.client;
+package com.joyent.manta.client.jobs;
 
+import com.joyent.manta.client.jobs.MantaJobPhase;
 import org.testng.annotations.Test;
 
 /**