+++ java-manta-it/src/test/java/com/joyent/manta/client/jobs/MantaClientJobIT.java
@@ -1,8 +1,9 @@
 /*
  * Copyright (c) 2015, Joyent, Inc. All rights reserved.
  */
-package com.joyent.manta.client;
+package com.joyent.manta.client.jobs;
 
+import com.joyent.manta.client.MantaClient;
 import com.joyent.manta.config.IntegrationTestConfigContext;
 import com.joyent.manta.config.ConfigContext;
 import com.joyent.test.util.ThreeTriesRetryAnalyzer;